requirements.txt
asgiref==3.5.0
autopep8==1.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
flake8==4.0.1
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
static3==0.7.0
toml==0.10.2
tzdata==2022.1
